fix observe section, fix index
