+++ components/Error.vue
@@ -4,7 +4,7 @@
       .error__message
         p {{errorMessage}}
         p.error__back
-          NuxtLink(:to="{name: 'index'}" v-t="'error.back'")
+          NuxtLink(:to="{name: 'index___' + $i18n.locale}" v-t="'error.back'")
       .error__cats
         img.error__cat(v-for="cat in limitCats" :key="cat.id" :src="cat.src" :style="cat.position")
     SocialHead(:title="errorTitle" :description="error.message")
@@ -27,6 +27,22 @@ export default Vue.extend({
       count: 0,
     };
   },
+  async fetch() {
+    const url = `https://cataas.com/api/cats?limit=50&tag=gif&skip=${
+      Math.random() * 100
+    }`;
+    const res = await this.$nuxt.context.$http.get(url);
+    const json = (await res.json()) as any[];
+    this.cats = json.map((item) => {
+      const top = Math.floor(Math.random() * 100);
+      const left = Math.floor(Math.random() * 100);
+      return {
+        src: `https://cataas.com/cat/${item.id}`,
+        position: `top:${top.toString()}%;left:${left.toString()}%;`,
+        id: item.id,
+      };
+    });
+  },
   methods: {
     addCat() {
       if (this.count > 50) {
@@ -34,23 +50,6 @@ export default Vue.extend({
         return;
       }
       this.count++;
-    },
-    async getCats() {
-      const url = `https://cataas.com/api/cats?limit=50&tag=gif&skip=${
-        Math.random() * 100
-      }`;
-      const res = await this.$nuxt.context.$http.get(url);
-      const json = (await res.json()) as any[];
-      const cats: Cat[] = json.map((item) => {
-        const top = Math.floor(Math.random() * 100);
-        const left = Math.floor(Math.random() * 100);
-        return {
-          src: `https://cataas.com/cat/${item.id}`,
-          position: `top:${top.toString()}%;left:${left.toString()}%;`,
-          id: item.id,
-        };
-      });
-      return cats;
     },
   },
   computed: {
@@ -70,11 +69,8 @@ export default Vue.extend({
     },
   },
   mounted() {
-    this.getCats().then((cats) => {
-      this.cats = cats;
-      this.addCat();
-      this.interval = window.setInterval(this.addCat, 1000);
-    });
+    this.addCat();
+    this.interval = window.setInterval(this.addCat, 1000);
   },
   beforeDestroy() {
     clearInterval(this.interval);
