+++ components/Header.vue
@@ -2,7 +2,7 @@
   header.header#header
     .header__container
       h1.header__title
-        NuxtLink(:to="{name: 'index'}" v-t="'header.title'")
+        NuxtLink(:to="{name: 'index___' + $i18n.locale}" v-t="'header.title'")
       nav
         h2(v-t="'header.nav.title'").is-hidden
         ul.header__nav
@@ -19,11 +19,6 @@ import type { LocaleObject } from 'nuxt-i18n';
 import slugify from 'slugify';
 
 export default Vue.extend({
-  data() {
-    return {
-      observer: null as null | IntersectionObserver,
-    };
-  },
   computed: {
     getSwitchLocalCode(): string | false {
       const locales = this.$i18n.locales as LocaleObject[];
@@ -35,64 +30,11 @@ export default Vue.extend({
   },
   methods: {
     getItemLink(item: string) {
-      return { name: 'index', hash: '#' + slugify(item, { lower: true }) };
+      return {
+        name: 'index___' + this.$i18n.locale,
+        hash: '#' + slugify(item, { lower: true }),
+      };
     },
-  },
-  beforeDestroy() {
-    this.observer?.disconnect();
-  },
-  mounted() {
-    this.observer = new IntersectionObserver(
-      (entries) => {
-        let intersectionRatio = 0;
-        let hash: string | false = false;
-
-        for (let index = 0; index < entries.length; index++) {
-          const entry = entries[index];
-          const intersectionEl = entry.target as HTMLElement;
-          const section = intersectionEl.closest('.section') as HTMLElement;
-
-          // update active
-          const hrefTitle = section?.querySelector('.section__title');
-          if (hrefTitle)
-            hrefTitle.classList[entry.isIntersecting ? 'add' : 'remove'](
-              'is-active'
-            );
-
-          // update anchor
-          if (
-            !entry.isIntersecting ||
-            intersectionRatio > entry.intersectionRatio
-          )
-            continue;
-          intersectionRatio = entry.intersectionRatio;
-
-          if (section && !section.classList.contains('section--about')) {
-            // The link
-            const id = section.id;
-            if (!id) return;
-            hash = id;
-          } else {
-            hash = '';
-          }
-        }
-
-        if (hash !== false) {
-          const url = new URL(window.location.toString());
-          url.hash = hash;
-          window.history.pushState(null, '', url.toString());
-        }
-      },
-      { threshold: 0 }
-    );
-
-    const items = Array.from(
-      document.getElementsByClassName('section__intersection')
-    );
-    for (let index = 0; index < items.length; index++) {
-      const item = items[index];
-      this.observer.observe(item);
-    }
   },
 });
 </script>
